+++ packages/terra-graphs-docs/CHANGELOG.md
@@ -22,3 +22,4 @@
 
 * Changed 
   * Moved core documentation to terra-dev-site.
+  * Updated Bubble graph documentation and examples to be carbon centric.